7-IMPROVEMENT Moved data_path from runs.ts to config.ts
