+++ main-module/src/runs.ts
@@ -5,6 +5,7 @@ import {HwConfig} from './homework'
 import path from 'path'
 import {StudentList} from "classroom-api";
 import {config} from "./config";
+import { data_path} from "./config";
 
 export interface RunOpts {
     trial?: boolean,
@@ -21,7 +22,7 @@ export function log<T>(e: T, message: string) {
     }
     return e
 }
-export const data_path = path.resolve(__dirname, `../../../data`)
+
 const results_path = `${data_path}/output`
 const submissions_path = `${data_path}/submissions`
