+++ main-module/src/scripts/sumResults.ts
@@ -1,6 +1,6 @@
 import fs from 'fs'
 import path from 'path'
-import {data_path} from '../runs'
+import {data_path} from '../config'
 import {DEFAULT_HW_CONFIG_PATH, readHomeworkConfiguration} from '../homework'
 import {mergeResults} from '../partitions'
